server/requirements.txt
flask>=2.0.0
flask-socketio>=5.0.0
python-dotenv>=0.19.0